syskit: implement support for models-in-models
This provides full support for having models in models, and making
it as transparent as possible within Syskit. In practice, the main
goal of this commit is to make sure that if one does e.g.

~~~
use_sdf_model 'model://ur10'
~~~

Then the profile and all it uses remains valid even if the ur10 model
is "moved" within the SDF hierarchy (as e.g. mounted on a mobile
platform).

In order to achieve this, two things need to be done:
 - the user may provide a name that is expected to remain consistent
   across the SDF scenes (e.g. 'arm'). This is used to locate the model
   in the overall scene.
 - the generated link and sensor devices need to be prefixed by the model
   name, so as to avoid collisions - which would become likely under the
   assumption that we 'move' the SDF tree around. The current behavior
   is retained by default, but a (loud) deprecation warning is issued.