syskit: allow already defined devices in separate profiles
There is a one-profile, one-model rule (because then a profile represents
a given robot/model, as intended). So, if a given system wants to
control more than one device, it must create more than one profile
and call use_gazebo_model separately.

This leads to having devices defined multiple times. Because the
merging algorithm looks at argument values, the corresponding ModelTask
are not merged and the deployment fails.

Introduce a `reuse:` flag that allows to specify that a given profile
should reuse devices from another, to avoid this.