support explicitly providing a task model to deploy for plugin blocks
Currently, the Rock tasks that are providing an interface for plugins
(thruster and underwater) had to live in the rock_gazebo orogen project,
and were hardcoded in the Ruby code.

This changes that, and allows to instead specify the task model
in a <task ... /> element. The task model can be arbitrary, and does
not therefore have to live in the rock_gazebo orogen model.

Additionally, this allows to reuse the same task for different plugins
(provided they have the same internal way to interact with the task)

The current behavior is retained for the underwater and thruster plugin,
as long as they do not have a <task ...>. If the task element is present,
the hardcoded code path is not taken.