fix(test): attempt to fix flaky test
I'm randomly (and rarely) getting ComError in CI during
configuration of some sensor component - not always the
same one.

Publish the task on CORBA only after its activity is set.
It's a long shot - I have no idea whether it's the source
of the problem - but in any case it's better. Activities
should really not be changed while the task is being
manipulated.