fix: make sure use_gazebo_model does not need use_gazebo_world to function
It seems that `use_gazebo_world` has been added to `use_gazebo_model` one day as
a workaround for problems, probably related to hierarchical models.

This commit actually fixes that problem, making sure we do *not* need to call
use_gazebo_world - which has a very high cost in complex scenes, and pollutes
the transformer with a lot of unneeded frames.

The main issue was that the root model frame was not automatically added to the
transformer, but it is needed to actually run the ModelTask.