fix: do not assume the transformer cannot be already enabled
The transformer is now complaining when Syskit.enable is called twice (this is
actually a bug). Unfortunately, rock_gazebo/syskit enables the transformer
"manually" - which it should really not do - and things fail if code enabled
thet transformer explicitly before require'ing rock_gazebo/syskit.

Make sure that this is allowed and a no-op